Install baselayout rather than duplicating dir/symlink behaviour here
crossdev environments need to support merged-usr now, especially now
that the profile can be selected. The logic around this is messy and
best handled by the baselayout package when it is supported by the arch.
This package now gets installed immediately before binutils. It means a
few small extra files get installed, but that's a small price to pay. We
could INSTALL_MASK them, but that would just add more complexity.

Signed-off-by: James Le Cuirot <chewi@gentoo.org>